Methods to serialize configuration objects into hashes

+++ app/models/form_section.rb
@@ -1,9 +1,10 @@
-class FormSection < ApplicationRecord
+# frozen_string_literal: true
 
+# A form is a collection of fields. Forms describe how a user may interact with a record.
+class FormSection < ApplicationRecord
   include LocalizableJsonProperty
   include ConfigurationRecord
   # include Importable # TODO: This will need to be rewritten
-  # include Memoizable
 
   RECORD_TYPES = %w[case incident tracing_request].freeze
 
@@ -34,13 +35,13 @@ class FormSection < ApplicationRecord
   end
 
   def generate_unique_id
-    if self.name_en.present? && self.unique_id.blank?
-      self.unique_id = self.name_en.gsub(/[^A-Za-z0-9_ ]/, '').parameterize.underscore
-    end
+    return unless name_en.present? && unique_id.blank?
+
+    self.unique_id = name_en.gsub(/[^A-Za-z0-9_ ]/, '').parameterize.underscore
   end
 
   # TODO: This method will go away after UIUX refactor
-  def form_group_name(opts={})
+  def form_group_name(opts = {})
     locale = (opts[:locale].present? ? opts[:locale] : I18n.locale)
     return name(locale) if form_group_id.blank?
 
@@ -54,30 +55,21 @@ class FormSection < ApplicationRecord
     Lookup.form_group_name_all(form_group_id, parent_form, module_name, lookups)
   end
 
-  def localized_property_hash(locale=Primero::Application::BASE_LANGUAGE, show_hidden_fields=false)
+  def localized_property_hash(locale = Primero::Application::BASE_LANGUAGE, show_hidden_fields = false)
     lh = localized_hash(locale)
     fldz = {}
-    self.fields.each { |f| fldz[f.name] = f.localized_property_hash locale if (show_hidden_fields || f.visible?)}
+    fields.each { |f| fldz[f.name] = f.localized_property_hash locale if show_hidden_fields || f.visible? }
     lh['fields'] = fldz
     lh
   end
 
   def inspect
-    "FormSection(#{self.name}, form_group_id => '#{self.form_group_id}')"
+    "FormSection(#{name}, form_group_id => '#{form_group_id}')"
   end
 
   alias_attribute :to_param, :unique_id
 
   class << self
-
-    def all_filterable_field_names(foo)
-      []
-    end
-
-    def memoized_dependencies
-      [Field]
-    end
-
     def permitted_api_params
       [
         'id', 'unique_id', { 'name' => {} }, { 'help_text' => {} }, { 'description' => {} }, 'parent_form',
@@ -94,36 +86,35 @@ class FormSection < ApplicationRecord
       form_section
     end
 
-    #TODO: Used by importer. Refactor?
+    # TODO: Used by importer. Refactor?
     def get_unique_instance(attributes)
       find_by(unique_id: attributes['unique_id'])
     end
-    #memoize_in_prod :get_unique_instance
+    # memoize_in_prod :get_unique_instance
 
-    #Given a list of forms, return their subforms
+    # Given a list of forms, return their subforms
     def get_subforms(forms)
       form_ids = forms.map(&:id)
       subform_fields = Field.includes(:subform).where(form_section_id: form_ids, type: Field::SUBFORM)
       subform_fields.map(&:subform).compact
     end
-    #memoize_in_prod :get_subforms
 
     def all_forms_grouped_by_parent(include_subforms=false)
       forms = FormSection.where(is_nested: false)
       forms = forms.unscope(:where) if include_subforms
-      forms.group_by{|f| f.parent_form}
+      forms.group_by { |f| f.parent_form }
     end
-    #memoize_in_prod :all_forms_grouped_by_parent
 
-    #Create the form section if does not exists.
-    #If the form section does exist will attempt
-    #to create fields if the fields does not exists.
+    # Create the form section if does not exists.
+    # If the form section does exist will attempt
+    # to create fields if the fields does not exists.
     def create_or_update_form_section(properties = {})
       unique_id = properties[:unique_id]
       return nil if unique_id.blank?
+
       fields = properties[:fields]
 
-      form_section = self.find_by(unique_id: unique_id)
+      form_section = find_by(unique_id: unique_id)
       if form_section.present?
         Rails.logger.info {"Updating form section #{unique_id}"}
         fields = fields.map do |field|
@@ -134,31 +125,29 @@ class FormSection < ApplicationRecord
         end
         properties[:fields] = fields
         form_section.update_attributes(properties)
-        return form_section
+        form_section
       else
         Rails.logger.info {"Creating form section #{unique_id}"}
-        return FormSection.create!(properties)
+        FormSection.create!(properties)
       end
     end
-    alias :create_or_update :create_or_update_form_section
+    alias create_or_update create_or_update_form_section
 
-    #TODO: This method may be removed, depending on how the all_searchable_* get refactored
+    # TODO: This method may be removed, depending on how the all_searchable_* get refactored
     def find_by_parent_form(parent_form, subforms=true)
       forms = FormSection.where(parent_form: parent_form, is_nested: false)
       forms = forms.unscope(where: :is_nested) if subforms
       forms.order(order_form_group: :asc, order: :asc, order_subform: :asc)
     end
-    #memoize_in_prod :find_by_parent_form
 
     def violation_forms
-      #TODO: Fix this when we make MRM work
+      # TODO: Fix this when we make MRM work
       # ids = Incident.violation_id_fields.keys
       # FormSection.by_unique_id(keys: ids).all
       []
     end
-    #memoize_in_prod :violation_forms
 
-    #TODO: This needs to be made not hard-coded. Used only in Exporters to exclude binary data
+    # TODO: This needs to be made not hard-coded. Used only in Exporters to exclude binary data
     def binary_form_names
       ['Photos and Audio', 'Other Documents', 'BID Records', 'BIA Records']
     end
@@ -167,30 +156,28 @@ class FormSection < ApplicationRecord
       Lookup.where("unique_id like 'lookup-form-group-%'")
     end
 
-    #Force eager loading of subforms
+    # Force eager loading of subforms
     def link_subforms(forms)
-      subform_fields = forms.map(&:fields).flatten.select{|f| f.type == Field::SUBFORM}
+      subform_fields = forms.map(&:fields).flatten.select { |f| f.type == Field::SUBFORM }
       ActiveRecord::Associations::Preloader.new.preload(subform_fields, :subform)
       subforms = subform_fields.map(&:subform)
-      ActiveRecord::Associations::Preloader.new.preload(subforms, [:fields, :collapsed_fields])
-      return forms
+      ActiveRecord::Associations::Preloader.new.preload(subforms, %i[fields collapsed_fields])
+      forms
     end
-    #memoize_in_prod :link_subforms
 
-    #Return a hash of subforms, where the keys are the form groupings
+    # Return a hash of subforms, where the keys are the form groupings
     # TODO: This method might no longer be relevant. Investigate! We should avoid sorting in Ruby. UIUX?
     def group_forms(forms)
       grouped_forms = {}
 
-      #Order these forms by group and form
-      sorted_forms = forms.sort_by{|f| [f.order_form_group, f.order]}
+      # Order these forms by group and form
+      sorted_forms = forms.sort_by { |f| [f.order_form_group, f.order] }
 
       if sorted_forms.present?
-        grouped_forms = sorted_forms.group_by{|f| f.form_group_name}
+        grouped_forms = sorted_forms.group_by { |f| f.form_group_name }
       end
-      return grouped_forms
+      grouped_forms
     end
-    #memoize_in_prod :group_forms
 
     def form_sections_by_ids_and_parent_form(form_ids, parent_form)
       FormSection.find_by_parent_form(parent_form, true).where(unique_id: form_ids)
@@ -198,6 +185,7 @@ class FormSection < ApplicationRecord
 
     def add_field_to_formsection(formsection, field)
       raise I18n.t("errors.models.form_section.add_field_to_form_section") unless formsection.editable
+
       field.form_section = formsection
       if field.type == Field::SUBFORM
         subform = create_subform(formsection, field)
@@ -207,32 +195,31 @@ class FormSection < ApplicationRecord
     end
 
     def create_subform(formsection, field)
-      self.create_or_update_form_section({
-                visible: false,
-                is_nested: true,
-                core_form: false,
-                editable: true,
-                order_form_group: formsection.order_form_group,
-                order: formsection.order,
-                order_subform: 1,
-                unique_id: field.name,
-                parent_form: formsection.parent_form,
-                name_all: field.display_name,  #TODO: is _all or _en better?
-                description_all: field.display_name
-      })
+      create_or_update_form_section(
+        visible: false,
+        is_nested: true,
+        core_form: false,
+        editable: true,
+        order_form_group: formsection.order_form_group,
+        order: formsection.order,
+        order_subform: 1,
+        unique_id: field.name,
+        parent_form: formsection.parent_form,
+        name_all: field.display_name, # TODO: is _all or _en better?
+        description_all: field.display_name
+      )
     end
 
     def has_photo_form
       photo_form = find_by(unique_id: 'photos_and_audio')
       photo_form.present? && photo_form.visible
     end
-    # memoize_in_prod :has_photo_form
 
-    def new_custom(form_section, module_name = "CP")
-      form_section[:core_form] = false   #Indicates this is a user-added form
+    def new_custom(form_section, module_name = 'CP')
+      form_section[:core_form] = false   # Indicates this is a user-added form
 
-      #TODO - need more elegant way to set the form's order
-      #form_section[:order] = by_order.last ? (by_order.last.order + 1) : 1
+      # TODO - need more elegant way to set the form's order
+      # form_section[:order] = by_order.last ? (by_order.last.order + 1) : 1
       form_section[:order] = 999
       form_section[:order_form_group] = 999
       form_section[:order_subform] = 0
@@ -240,7 +227,7 @@ class FormSection < ApplicationRecord
 
       fs = FormSection.new(form_section)
       fs.unique_id = "#{module_name}_#{fs.name_en}".parameterize.underscore
-      return fs
+      fs
     end
 
     def determine_parent_form(record_type, apply_to_reports=false)
@@ -275,22 +262,25 @@ class FormSection < ApplicationRecord
       end
     end
 
+    # TODO: Review; this method might no longer be relevant
     alias super_clear clear
     def clear
       Field.delete_all
-      self.all.each do |f|
+      all.each do |f|
         f.roles.destroy(f.roles)
       end
       super_clear
     end
 
+    # TODO: Review; this method might no longer be relevant
     def import(data)
-      form = self.create!(data.except('fields'))
+      form = create!(data.except('fields'))
       Field.import(data['fields'], form)
     end
 
+    # TODO: Review; this method might no longer be relevant
     def export
-      self.all.map do |record|
+      all.map do |record|
         record.attributes.tap do |form|
           form.delete('id')
           form['fields'] = record.fields.map(&:export)
@@ -298,28 +288,26 @@ class FormSection < ApplicationRecord
       end
     end
 
-    def list(params={})
+    def list(params = {})
       form_sections = all.includes(:fields, :collapsed_fields, :primero_modules)
       form_sections = form_sections.where(parent_form: params[:record_type]) if params[:record_type]
       form_sections = form_sections.where(primero_modules: { unique_id: params[:module_id] }) if params[:module_id]
       form_sections
     end
-
-
   end
 
   def all_mobile_fields
-    self.fields.select{|f| f.is_mobile?}
+    fields.select{|f| f.is_mobile?}
   end
 
   def localized_attributes_hash(locales)
     attributes = self.attributes.clone
-    #convert top level attributes
+    # convert top level attributes
     FormSection.localized_properties.each do |property|
       attributes[property] = {}
-      key = "#{property.to_s}_i18n"
-      Primero::Application::locales.each do |locale|
-        value = attributes.try(:[], key).try(:[], locale) || ""
+      key = "#{property}_i18n"
+      Primero::Application.locales.each do |locale|
+        value = attributes.try(:[], key).try(:[], locale) || ''
         attributes[property][locale] = value if locales.include? locale
       end
       attributes.delete(key)
@@ -327,14 +315,21 @@ class FormSection < ApplicationRecord
     attributes
   end
 
-  #TODO: Do we still need this after splitting FormSection and Field?
-  def properties= properties
+  def configuration_hash
+    hash = attributes.except('id')
+    hash['fields'] = fields.map(&:configuration_hash)
+    hash['module_unique_ids'] = primero_modules.pluck(:unique_id)
+    hash.with_indifferent_access
+  end
+
+  # TODO: Do we still need this after splitting FormSection and Field?
+  def properties=(properties)
     properties.each_pair do |name, value|
-      self.send("#{name}=", value) unless value == nil
+      send("#{name}=", value) unless value == nil
     end
   end
 
-  #TODO: Refactor with Field. No longer necessary?
+  # TODO: Refactor with Field. No longer necessary?
   # def add_field field
   #   self["fields"] << Field.new(field)
   # end
@@ -343,13 +338,12 @@ class FormSection < ApplicationRecord
     unique_id
   end
 
+  # TODO: Is this used?
   def delete_field(field_to_delete)
-    field = self.fields.where(name: field_to_delete, editable: true).first
-    if (field)
-      field.destroy
-    else
-      raise I18n.t("errors.models.form_section.delete_field")
-    end
+    field = fields.where(name: field_to_delete, editable: true).first
+    return field.destroy if field
+
+    raise I18n.t("errors.models.form_section.delete_field")
   end
 
   #TODO: Refactor or deprecate when we change UIUX
@@ -449,6 +443,17 @@ class FormSection < ApplicationRecord
     end
   end
 
+  # Location::LIMIT_FOR_API is a hard limit
+  # The SystemSettings limit is a soft limit that lets us adjust down below the hard limit if necessary
+  # #TODO: Refactor with Location
+  def self.include_locations_for_mobile?
+    location_count = Location.count
+    return false if location_count > Location::LIMIT_FOR_API
+
+    ss_limit = SystemSettings.current.try(:location_limit_for_api)
+    ss_limit.present? ? location_count < ss_limit : true
+  end
+
   protected
 
   def recalculate_collapsed_fields
@@ -539,23 +544,12 @@ class FormSection < ApplicationRecord
 
   private
 
-  # Location::LIMIT_FOR_API is a hard limit
-  # The SystemSettings limit is a soft limit that lets us adjust down below the hard limit if necessary
-  # #TODO: Refactor with Location
-  def self.include_locations_for_mobile?
-    location_count = Location.count
-    return false if location_count > Location::LIMIT_FOR_API
-    ss_limit = SystemSettings.current.try(:location_limit_for_api)
-    return_value = ss_limit.present? ? location_count < ss_limit : true
-  end
-
-  def update_field_translations(fields_hash={}, locale)
+  def update_field_translations(fields_hash = {}, locale)
     fields_hash.each do |key, value|
-      field = Field.find_by(name: key, form_section_id: self.id)
+      field = Field.find_by(name: key, form_section_id: id)
       if field.present?
         field.update_translations(value, locale)
       end
     end
   end
-
 end
